ATTN UdS ! Resolves Issue 81. Resolves Issue 80.
Added an InitServlet that is configured on deployment (as specified in web.xml). The InitServlet contains the ServletContext and sets it statically so as other classes can call it and retrieve the real path of files. This also allows for initialliasing log4j properly through the log4j.properties file. Note change in GeneralUtil to retrieve the real path. 

Hope that helps and it works on your side. Please test and we can discuss improvements.